Created main solution and C# wrapper library
Because AHK is written in C++, and I wanted it to be able to work with
non-.NET languages (like Python, JavaScript using NW.js/Electron etc.)
I figured the main DLL should be done in C++.  And eventually I'd like
to see wrappers in every language that'll run a DLL on Windows, but for
now I figured C# would be a pretty good start (and maybe Python next).

I also got the 2 easiest built-in variables, A_ScreenWidth/Height, done.